Added abstract has_weights, has_bias properties
